# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The transactional core: agreements, orders, the shipment lifecycle and invoices.\\n\\nStatus changes go through state machines rather than accepting an arbitrary status string, so an illegal transition is refused with 409 rather than written.",
        "title": "Karlo Business API",
        "termsOfService": "https://karlo.co.id/terms",
        "contact": {
            "name": "Karlo Engineering",
            "email": "engineering@karlo.co.id"
        },
        "version": "1.0"
    },
    "host": "localhost:5003",
    "basePath": "/api/v1",
    "paths": {
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "List orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_karlo_business-service_internal_platform_response.Meta"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create order",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_karlo_business-service_internal_models.Order"
                        }
                    }
                }
            }
        },
        "/orders/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Order dashboard summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_karlo_business-service_internal_models.Order"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Update a draft order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/orders/{id}/assign": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Assign a driver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_karlo_business-service_internal_models.Order"
                        }
                    }
                }
            }
        },
        "/orders/{id}/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Order status history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/orders/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Change order status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_karlo_business-service_internal_models.Order"
                        }
                    }
                }
            }
        },
        "/orders/{id}/transitions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Available order transitions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_karlo_business-service_internal_models.JSONB": {
            "type": "object",
            "additionalProperties": true
        },
        "github_com_karlo_business-service_internal_models.Order": {
            "type": "object",
            "properties": {
                "agreementId": {
                    "type": "string"
                },
                "cancelReason": {
                    "type": "string"
                },
                "cancelledAt": {
                    "type": "string"
                },
                "cancelledByUserId": {
                    "type": "string"
                },
                "cargoTypeId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdByUserId": {
                    "type": "string"
                },
                "currencyId": {
                    "type": "string"
                },
                "customerId": {
                    "type": "string"
                },
                "deliveryAt": {
                    "type": "string"
                },
                "destinationWarehouseId": {
                    "type": "string"
                },
                "detail": {
                    "$ref": "#/definitions/github_com_karlo_business-service_internal_models.JSONB"
                },
                "driverUserId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "itemTypeId": {
                    "type": "string"
                },
                "legacyId": {
                    "type": "string"
                },
                "orderKind": {
                    "type": "string"
                },
                "orderNumber": {
                    "type": "string"
                },
                "originWarehouseId": {
                    "type": "string"
                },
                "parentOrderId": {
                    "type": "string"
                },
                "pickupAt": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "number"
                },
                "referenceNumber": {
                    "type": "string"
                },
                "shipperCompanyId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "statusAlias": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "string"
                },
                "transporterCompanyId": {
                    "type": "string"
                },
                "truckId": {
                    "description": "TruckID is a Mongo ObjectId hex string, owned by the master data service.",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "volumeM3": {
                    "type": "number"
                },
                "weightKg": {
                    "type": "number"
                }
            }
        },
...
        "github_com_karlo_business-service_internal_platform_response.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRows": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "RS256 access token issued by the authentication service, as \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}